+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Jest/RTL: Cover `useOptions` hook with unit tests. Refs UIPCIR-68.
 * Jest/RTL: Cover `CreateRecordsWrapper` component with unit tests. Refs UIPCIR-64.
 * Jest/RTL: Cover `CreateRecordsForm` component with unit tests. Refs UIPCIR-66.
+* Jest/RTL: Cover `CreateRecordsPlugin` component with unit tests. Refs UIPCIR-65.
 
 ## [4.0.0](https://github.com/folio-org/ui-plugin-create-inventory-records/tree/v4.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-plugin-create-inventory-records/compare/v3.3.0...v4.0.0)